+++ genium/src/main/java/com/here/genium/generator/swift/SwiftArrayGenerator.java
@@ -38,14 +38,14 @@ public final class SwiftArrayGenerator {
   }
 
   public List<GeneratedFile> generate() {
-    SwiftFile arrayFile = new SwiftFile();
+    SwiftFile arrayFile = new SwiftFile(SWIFT_ARRAY);
     arrayFile.arrays.addAll(new TreeMap<>(arrayCollector).values());
 
     if (arrayFile.isEmpty()) {
       return emptyList();
     } else {
       String content = TemplateEngine.INSTANCE.render("swift/Array", arrayFile);
-      return Collections.singletonList(new GeneratedFile(content, SWIFT_ARRAY));
+      return Collections.singletonList(new GeneratedFile(content, arrayFile.fileName));
     }
   }
 }