+++ genium/src/main/java/com/here/genium/model/swift/SwiftClass.kt
@@ -19,6 +19,9 @@
 
 package com.here.genium.model.swift
 
+import java.util.function.Function
+import java.util.stream.Stream
+
 class SwiftClass(
     name: String,
     visibility: SwiftVisibility? = null,
@@ -50,4 +53,18 @@ class SwiftClass(
     @Suppress("unused")
     val constructors
         get() = methods.filter { it.isConstructor }
+
+    override fun stream(): Stream<SwiftModelElement>? {
+        return Stream.of(
+            super.stream(),
+            properties.stream(),
+            methods.stream(),
+            structs.stream(),
+            enums.stream(),
+            typedefs.stream(),
+            constants.stream()
+        ).flatMap(
+            Function.identity()
+        )
+    }
 }