+++ genium/src/main/java/com/here/genium/model/swift/SwiftStruct.kt
@@ -19,6 +19,9 @@
 
 package com.here.genium.model.swift
 
+import java.util.function.Function
+import java.util.stream.Stream
+
 class SwiftStruct(
     name: String,
     cPrefix: String = "",
@@ -78,4 +81,8 @@ class SwiftStruct(
 
         return swiftStruct
     }
+
+    override fun stream(): Stream<SwiftModelElement>? =
+        Stream.of(super.stream(), fields.stream(), constants.stream()).flatMap(
+            Function.identity())
 }