+++ genium/src/main/java/com/here/genium/platform/swift/SwiftGeneratorSuite.kt
@@ -22,11 +22,15 @@ package com.here.genium.platform.swift
 import com.here.genium.Genium
 import com.here.genium.generator.cbridge.CBridgeGenerator
 import com.here.genium.generator.common.GeneratedFile
+import com.here.genium.generator.common.templates.TemplateEngine
 import com.here.genium.generator.cpp.CppIncludeResolver
 import com.here.genium.generator.cpp.CppNameResolver
 import com.here.genium.generator.swift.SwiftGenerator
+import com.here.genium.generator.swift.SwiftModel
 import com.here.genium.model.cbridge.CBridgeIncludeResolver
 import com.here.genium.model.lime.LimeModel
+import com.here.genium.model.swift.SwiftMethod
+import com.here.genium.model.swift.SwiftModelElement
 import com.here.genium.platform.common.GeneratorSuite
 
 /**
@@ -38,6 +42,7 @@ import com.here.genium.platform.common.GeneratorSuite
 class SwiftGeneratorSuite(options: Genium.Options) : GeneratorSuite() {
     private val internalNamespace = options.cppInternalNamespace ?: emptyList()
     private val rootNamespace = options.cppRootNamespace
+    private val commentsProcessor = SwiftCommentsProcessor()
 
     override fun generate(limeModel: LimeModel): List<GeneratedFile> {
         val limeReferenceMap = limeModel.referenceMap
@@ -49,8 +54,45 @@ class SwiftGeneratorSuite(options: Genium.Options) : GeneratorSuite() {
             CppNameResolver(rootNamespace, limeReferenceMap),
             internalNamespace
         )
+        val swiftModel = limeModel.containers.fold(SwiftModel(emptyMap(), emptySet())) { model, limeContainer ->
+            model.merge(swiftGenerator.generateModel(limeContainer))
+        }
 
-        val result = limeModel.containers.map { swiftGenerator.generate(it) } +
+        val elementToSwiftName = mutableMapOf<SwiftModelElement, String>()
+        fun resolveFullName(element: SwiftModelElement, name: String) {
+            var elementName = element.simpleName
+            if (element is SwiftMethod) {
+                elementName = elementName + "(...)"
+            }
+            val fullName = if (name.isEmpty()) elementName else name + "." + elementName
+
+            element.stream().forEach { resolveFullName(it, fullName) }
+
+            elementToSwiftName[element] = fullName
+        }
+        swiftModel.containers.forEach { resolveFullName(it, "") }
+
+        val limeToSwiftName = swiftModel.referenceMap.mapValues { elementToSwiftName[it.value] ?: "" }
+
+        swiftModel.containers.forEach { topLevelElement ->
+            topLevelElement.streamRecursive().forEach { element ->
+                if (element is SwiftModelElement) {
+                    val fullLimeName =
+                        swiftModel.referenceMap.entries.firstOrNull { it.value == element }
+                    if (fullLimeName != null) {
+                        element.comment = commentsProcessor.processLinks(
+                            fullLimeName.key,
+                            element.comment,
+                            limeToSwiftName
+                        )
+                    }
+                }
+            }
+        }
+
+        val result = swiftModel.containers.filter { !it.isEmpty }.map { GeneratedFile(
+                TemplateEngine.render("swift/File", it),
+                it.fileName) } +
                 limeModel.containers.flatMap { cBridgeGenerator.generate(it) } +
             CBridgeGenerator.STATIC_FILES + SwiftGenerator.STATIC_FILES +
             cBridgeGenerator.arrayGenerator.generate() + swiftGenerator.arrayGenerator.generate() +